+++ indigo_app/models.py
@@ -5,7 +5,7 @@ from django.dispatch import receiver
 
 from rest_framework.authtoken.models import Token
 
-from indigo_api.models import Country
+from indigo_api.models import Country, Locality, Work, Document
 
 
 class Editor(models.Model):
@@ -50,6 +50,42 @@ class Publication(models.Model):
 
     def __unicode__(self):
         return unicode(self.name)
+
+
+class Task(models.Model):
+    title = models.CharField(max_length=256, null=False, blank=False)
+    content = models.TextField(null=True, blank=True)
+
+    country = models.ForeignKey(Country, related_name='tasks', null=False, blank=False, on_delete=models.CASCADE)
+    locality = models.ForeignKey(Locality, related_name='tasks', null=True, blank=True, on_delete=models.CASCADE)
+    work = models.ForeignKey(Work, related_name='tasks', null=True, blank=True, on_delete=models.CASCADE)
+    document = models.ForeignKey(Document, related_name='tasks', null=True, blank=True, on_delete=models.CASCADE)
+
+    # cf indigo_api.models.Annotation
+    anchor_id = models.CharField(max_length=128, null=True, blank=True)
+
+    state = models.CharField(max_length=128, null=False, blank=False, choices=(
+        ('open', 'open'), ('cancelled', 'cancelled'), ('pending', 'pending review'), ('closed', 'done')
+    ), default='open')
+
+    assigned_to = models.ForeignKey(User, related_name='assigned_tasks', null=True, blank=True, on_delete=models.SET_NULL)
+
+    created_by = models.ForeignKey(User, related_name='+', null=True, on_delete=models.SET_NULL)
+    updated_by = models.ForeignKey(User, related_name='+', null=True, on_delete=models.SET_NULL)
+
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    @property
+    def place_code(self):
+        return self.country.code + '-' + self.locality.code if self.locality else self.country.code
+
+
+class Workflow(models.Model):
+    title = models.CharField(max_length=256, null=False, blank=False)
+    description = models.TextField(null=True, blank=True)
+
+    tasks = models.ManyToManyField(Task, related_name='workflows')
 
 
 @receiver(post_save, sender=User)
